+++ src/utils/utils.ts
@@ -11,3 +11,13 @@ export const isJsonArray = (arg: any): arg is JsonArray => {
 export const isJsonObject = (arg: any): arg is JsonObject => {
   return typeof arg === 'object' && !Array.isArray(arg);
 };
+
+export const html = (strings: TemplateStringsArray, ...keys: any) =>
+  strings
+    .map((str, i) => {
+      if (i < strings.length - 1) {
+        return str + keys[i];
+      }
+      return str;
+    })
+    .join('');