+++ CHANGELOG.md
@@ -4,7 +4,14 @@
 
 - Merge multiple config files
 - More conflict strategy options
+- Better looking merge webview
 - Testing
+
+## [0.2.0] - 2021-03-30
+
+### Added
+
+- Command palette option to merge configurations in a webview.
 
 ## [0.0.1] - 2021-03-26
 
@@ -16,4 +23,5 @@
 - Merge config file with other text configs
 
 [unreleased]: https://github.com/aboqasem/vscode-mergeconf/
+[0.2.0]: https://github.com/aboqasem/vscode-mergeconf/releases/tag/v0.0.1
 [0.0.1]: https://github.com/aboqasem/vscode-mergeconf/releases/tag/v0.0.1